requirements.txt
openai==1.35.7 
python-dotenv==1.0.1 
fastapi==0.115.2 
uvicorn==0.32.0 
jinja2==3.1.4 
pytest==8.3.2
httpx==0.27.0